Fail requests when their responsible worker gets garbage collected
This is a best-effort mechanism to encourage waiters to unblock themselves
if the worker responsible for unblocking them exits before it does so.

The exact behavior here relies on implementation details of Go's garbage
collector and so may change in future, but as long as the runtime has
something resembling a garbage collector that is able to collect a
Worker object relatively soon after it's been dropped this should do the
job.