A sync.Once-like helper that is workgraph-aware
This is far more convenient in simple situations where there's no need to
delegate requests between workers: the request and associated promise are
entirely hidden and the caller just deals with a function that blocks
until the result is ready.

This also includes a further helper OnceFunc, which is similar to
sync.OnceValues in that it encapsulates the run-only-once function inside
its result so that it can be passed around to be called from various
locations without any need to know exactly what inner function it's
running.

When using these helpers callers need only worry about Worker objects,
which retain their purpose as our means for detecting self-dependencies.